Tighten up the grammar and fix aliases.
Documents now consist of 3 sections, in order:
title
alias
sequence

With this change, it's now impossible to end up with an ambiguity when
a name is encountered in a signal. The name will either be an actor
name or an alias, but never both.

This change also fixes an issue where it was possible to have two
separate actors with the exact same name.